% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny App}
\usage{
run_app()
}
\value{
A Shiny app object.
}
\description{
This function launches the Shiny app for exploring how US federal taxes are distributed.
}
\examples{
if (interactive()) {
  usfederaltaxexplorer::run_app()
}
}
